🔁 refac for edge, server auth
