+++ src/app/lib/auth/constants.ts
@@ -1,0 +1,3 @@
+export const REASON_NO_TOKEN = "No token provided in 'Authorization' header";
+export const REASON_INVALID_BEARER_TOKEN = "Invalid Bearer Token";
+export const REASON_DISABLED_USER = "User disabled";
